cleanup code

+++ app/src/main/java/ru/bagrusss/revolutdemo/mappers/impl/RatesMapper.kt
@@ -9,9 +9,9 @@ import javax.inject.Inject
 /**
  * Created by bagrusss on 23.09.2019
  */
-class RatesMapper @Inject constructor(private val resourcesProvider: ResourcesProvider): Mapper<Map<String, BigDecimal>, List<Rate>> {
+class RatesMapper @Inject constructor(private val resourcesProvider: ResourcesProvider): Mapper<Map<String, Double>, List<Rate>> {
 
-    override fun map(from: Map<String, BigDecimal>): List<Rate> = from.map { (rate, cost) ->
+    override fun map(from: Map<String, Double>): List<Rate> = from.map { (rate, cost) ->
         val (description, img) = resourcesProvider.rateImageAndDescription(rate)
         Rate(
             title = rate,
