+++ app/src/main/java/ru/bagrusss/revolutdemo/rates/models/Rate.kt
@@ -1,7 +1,5 @@
 package ru.bagrusss.revolutdemo.rates.models
 
-import java.math.BigDecimal
-
 /**
  * Created by bagrusss on 13.08.2019
  */
@@ -10,5 +8,5 @@ data class Rate(
     @JvmField val title: String,
     @JvmField val description: String,
     @JvmField val imgUrl: String,
-    @JvmField val cost: BigDecimal
+    @JvmField val cost: Double
 )
