+++ app/src/main/java/ru/bagrusss/revolutdemo/repository/RatesRepository.kt
@@ -9,7 +9,7 @@ import java.math.BigDecimal
  * Created by bagrusss on 13.08.2019
  */
 interface RatesRepository {
-    var currentBaseRate: Pair<String, BigDecimal>
+    var currentBaseRate: Pair<String, Double>
     val currentCostChanges: Observable<List<Rate>>
     val actualRates: Single<List<Rate>>
 }
